requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.40.0
peft>=0.10.0
trl>=0.8.6
bitsandbytes>=0.43.0
accelerate>=0.30.0
datasets>=2.19.0

# Tokenization
sentencepiece>=0.2.0
tokenizers>=0.19.0

# Config
pyyaml>=6.0.0

# Logging
wandb>=0.17.0
